@@ -24,7 +24,7 @@ setup(
     author="Sebastian Pipping",
     author_email="sebastian@pipping.org",
     url=f"https://github.com/hartwork/{APP}",
-    python_requires=">=3.9",
+    python_requires=">=3.10",
     setup_requires=[
         "setuptools>=38.6.0",  # for long_description_content_type
     ],
@@ -48,7 +48,6 @@ setup(
         "Programming Language :: Python",
         "Programming Language :: Python :: 3",
         "Programming Language :: Python :: 3 :: Only",
-        "Programming Language :: Python :: 3.9",
         "Programming Language :: Python :: 3.10",
         "Programming Language :: Python :: 3.11",
         "Programming Language :: Python :: 3.12",
